(patch) Moving Logic to correct places
