+++ src/CampusLearn.Domain/Account/IAccountFactory.cs
@@ -8,7 +8,8 @@ namespace CampusLearn.Domain.Account
 {
     public interface IAccountFactory
     {
-        IAccount CreateStudent(string name, string id);
+        IAccount CreateStudent(string name, int id);
         IAccount CreateTutor(string name, string subject);
+        IAccount CreateAdmin(string name);
     }
 }
