+++ src/CampusLearn.Infrastructure/Account/Admin.cs
@@ -1,5 +1,4 @@
 ﻿using CampusLearn.Domain.Account;
-using CampusLearn.Domain.TopicsNS;
 
 namespace CampusLearn.Infrastructure.Account;
